fix: adapt to tree-sitter 0.27 node API changes
tree-sitter 0.27 makes two breaking changes that this bump did not account
for, so the branch failed to compile.

QueryMatch::captures is now a private field with a captures() accessor
returning the same slice, so the two iteration sites in flags.rs call it.

Node::child_count() now returns u32 rather than usize, and Node::child()
takes u32, so the 40 `as u32` casts that converted the old usize return are
dead. Removing them is mechanical and changes no behavior; clippy rejects
them under -D warnings as casts to the same type.

+++ src/analyzers/mutation/operators/python/identity.rs
@@ -64,7 +64,7 @@ impl PythonIdentityOperator {
         };
 
         // Look for specific operators in children
-        for i in 0..node.child_count() as u32 {
+        for i in 0..node.child_count() {
             let child = match node.child(i) {
                 Some(c) => c,
                 None => continue,
